pbsys: Split program load from user program.
Some builds support user programs, but not loading new programs.
