+++ bricks/stm32/stm32.mk
@@ -496,9 +496,9 @@ PBIO_SRC_C = $(addprefix lib/pbio/,\
 	sys/light_matrix.c \
 	sys/light.c \
 	sys/main.c \
+	sys/program_load.c \
 	sys/status.c \
 	sys/supervisor.c \
-	sys/system.c \
 	sys/user_program.c \
 	)
