+++ bricks/virtualhub/mpconfigvariant.mk
@@ -183,9 +183,10 @@ LIB_SRC_C += $(addprefix lib/,\
 	pbio/sys/io_ports.c \
 	pbio/sys/light_matrix.c \
 	pbio/sys/light.c \
+	pbio/sys/main.c \
+	pbio/sys/program_load.c \
 	pbio/sys/status.c \
 	pbio/sys/supervisor.c \
-	pbio/sys/system.c \
 	pbio/sys/user_program.c \
 	)
