+++ lib/pbio/include/pbsys/main.h
@@ -1,0 +1,11 @@
+// SPDX-License-Identifier: MIT
+// Copyright (c) 2018-2022 The Pybricks Authors
+
+#ifndef _PBSYS_MAIN_H_
+#define _PBSYS_MAIN_H_
+
+#include "pbsys/config.h"
+
+void pbsys_init(void);
+
+#endif // _PBSYS_MAIN_H_
