--- lib/pbio/include/pbsys/system.h
@@ -1,9 +1,0 @@
-// SPDX-License-Identifier: MIT
-// Copyright (c) 2022 The Pybricks Authors
-
-#ifndef _PBSYS_SYS_SYSTEM_H_
-#define _PBSYS_SYS_SYSTEM_H_
-
-void pbsys_init(void);
-
-#endif // _PBSYS_SYS_SYSTEM_H_
