+++ lib/pbio/platform/nxt/pbsysconfig.h
@@ -1,8 +1,8 @@
 // SPDX-License-Identifier: MIT
-// Copyright (c) 2020-2021 The Pybricks Authors
+// Copyright (c) 2020-2022 The Pybricks Authors
 
 #define PBSYS_CONFIG_BATTERY_CHARGER                (0)
 #define PBSYS_CONFIG_BLUETOOTH                      (0)
 #define PBSYS_CONFIG_HUB_LIGHT_MATRIX               (0)
 #define PBSYS_CONFIG_STATUS_LIGHT                   (0)
-#define PBSYS_CONFIG_USER_PROGRAM_LOAD              (0)
+#define PBSYS_CONFIG_PROGRAM_LOAD                   (0)
