+++ lib/pbio/platform/virtual_hub/pbsysconfig.h
@@ -5,4 +5,4 @@
 #define PBSYS_CONFIG_BLUETOOTH                      (0)
 #define PBSYS_CONFIG_HUB_LIGHT_MATRIX               (0)
 #define PBSYS_CONFIG_STATUS_LIGHT                   (1)
-#define PBSYS_CONFIG_USER_PROGRAM_LOAD              (0)
+#define PBSYS_CONFIG_PROGRAM_LOAD                   (0)
